DjangoQuiz/requirements.txt
#---------Backend---------------
python              ^= 3.10.7
Django              ^= 4.1.1
Django-nested-admin ^= 3.4.0
djangorestframewokr ^= 3.14.0
pip                 ^= 22.2.2
django-rest-knox    ^= 4.2.0
django-cors-headers ^= 3.13.0

#---------Frontend---------------
npm                 ^= 8.19.2
angular             ^= 14.2.3
nodejs              ^= 16.17.1
ngx-cookie-service  ^= 